requirements.txt
torch>=2.0
monai>=1.3
nibabel>=5.0
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.10
pyyaml>=6.0
tqdm>=4.65
matplotlib>=3.7
